Changes to accommodate Views.
Views contain only Trees and Leaves. We wanted data aggregation to work
for Views, since we the Data limb works for Trees. To make this work
well needed to make some adjustments to allow for the presence of Leaves
as well as to set the default aggregation by 'path' instead of 'uuid',
since only Treants have uuids.