Now deletion of file deletes proxy file, too.
Presence of proxy file was keeping directories from being deleted. This
is no longer a problem, but the solution is a bit hardcoded. If the
naming convention of proxy files changes (unlikely), then this will be
broken again. Again, this module needs a major refactor anyway, so this
is a bandaid.